Vitest init
