+++ package.json
@@ -9,7 +9,7 @@
   "scripts": {
     "start": "node index.js",
     "dev": "nodemon index.js",
-    "test": "jest"
+    "test": "vitest run"
   },
   "keywords": [],
   "author": "",
@@ -40,6 +40,8 @@
   },
   "devDependencies": {
     "nodemon": "^3.1.9",
-    "prisma": "^7.3.0"
+    "prisma": "^7.3.0",
+    "supertest": "^7.2.2",
+    "vitest": "^4.0.18"
   }
 }
